backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
pydantic==2.5.3
pydantic-settings==2.1.0
pyjwt==2.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
mangum==0.17.0
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0

# Phase III: AI Chatbot Dependencies
google-generativeai==0.3.2
mcp==1.1.2
